Standardise layout of specs
It turns out I'm not the only one that decided to turn the spec
into a markdown file. There is a pretty common markdown version
floating around. The wording is almost identical to this, however
the formatting chose was quite different making comparsion
difficult.

This updates the formatting of my spec so that it more closely
matches other versions I have seen.

I have also added the word "Example" in front of each example.