Add an efficient multistage Dockerfile
This Dockerfile was added to test bulding and runner docker containers
via Github actions. It splits things up into multiple stages so that the
final image is much smaller as it does not need all the tools which were
needed to build the gems that I have included in the Gemfile.